update time_downsample_rate calculation in scripts/keyframe*.py

+++ scripts/keyframe_arkit_bs_from_nv_a2f_csv_for_nv_glb.py
@@ -290,7 +290,7 @@ if __name__ == "__main__":
     target_fps = 30
     # time_downsample_rate = 4  # 100 fps -> 25 fps
     # time_downsample_rate = 2  # 60 fps -> 30 fps
-    time_downsample_rate = input_fps / target_fps
+    time_downsample_rate = int(input_fps / target_fps)
 
     for mesh_obj in arkit_rigged_mesh_obj_names:
         clear_keyframed_animation_data(mesh_obj)
